# docs/swagger.yaml
basePath: /
definitions:
  dependencymodel.BpmnResource:
    properties:
      id:
        type: string
    type: object
  dependencymodel.Dependencies:
    properties:
      deployment_id:
        description: from db
        type: string
      devices:
        items:
          $ref: '#/definitions/dependencymodel.DeviceDependency'
        type: array
      events:
        items:
          $ref: '#/definitions/dependencymodel.EventDependency'
        type: array
      owner:
        type: string
    type: object
  dependencymodel.DeviceDependency:
    properties:
      bpmn_resources:
        items:
          $ref: '#/definitions/dependencymodel.BpmnResource'
        type: array
      device_id:
        type: string
      name:
        type: string
    type: object
  dependencymodel.EventDependency:
    properties:
      bpmn_resources:
        items:
          $ref: '#/definitions/dependencymodel.BpmnResource'
        type: array
      event_id:
        type: string
    type: object
...
info:
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Process-Deployment API
  version: "0.1"
paths:
  /dependencies:
    get:
      description: list dependencies
      parameters:
      - description: default 100, will be ignored if 'ids' is set
        in: query
        name: limit
        type: integer
      - description: default 0, will be ignored if 'ids' is set
        in: query
        name: offset
        type: integer
      - description: filter; ignores limit/offset; comma-seperated list
        in: query
        name: ids
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dependencymodel.Dependencies'
            type: array
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: list dependencies
      tags:
      - dependencies
  /dependencies/{id}:
    get:
      description: get dependencies of deployment
      parameters:
      - description: deployment id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dependencymodel.Dependencies'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: get dependencies
      tags:
      - dependencies
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"